feat(verify): dispatch chain-verify K=5..8 onto the fused wyN pool arm
- extract the K=17 wy17 arm body into a shared pool-layout helper
  (trait_decode_batched_conv_gdn_wyn.rs) — fused conv_kn epilogue + one
  wyN launch — dispatched identically for K=17 and K=5..8
- rename gdn_decode_wy17 -> gdn_decode_wyn (the wrapper was already
  K-agnostic; the kernel handle selects K)
- gdn_wyn_k: [KernelHandle; 4] loaded via try_kernel; K=5..8 arm gates
  on handle presence + h-intermediates pool contiguity (fails safe to
  the sequential fallback), kill-switch ATLAS_GDN_WYN=0
- removes the serial per-token GDN fallback at chain-verify K in 5..8
