+++ crates/spark-model/src/layers/qwen3_ssm/trait_decode_batched_conv_gdn.rs
@@ -3,8 +3,10 @@
 //! Phase 5-7 of `Qwen3SsmLayer::decode_batched_inner`: Conv1d + L2 norm +
 //! GDN per-token (with intermediate checkpoints). Extracted from
 //! `trait_decode_batched.rs` to keep the parent file under 500 LoC.
-//! Dispatches one of the fused K=2/3/4/17 paths or the sequential
-//! per-token fallback. All buffers + state are owned by the caller; this
+//! Dispatches one of the fused K=2/3/4 paths, the pool-layout WY arm
+//! (K∈{5..8} chain verify and K=17 DFlash — see
+//! `trait_decode_batched_conv_gdn_wyn.rs`), or the sequential per-token
+//! fallback. All buffers + state are owned by the caller; this
 //! function only mutates `ssm_state.h_state`, `ssm_state.conv_state`,
 //! their intermediates, `conv_out_buf`, and `gdn_out_buf`.
 
@@ -66,8 +68,9 @@ impl Qwen3SsmLayer {
     }
 
     /// Run conv1d_update_l2norm + GDN over `num_tokens` (multi-token decode
-    /// / MTP verify). Picks the K=2/3/4/17 fused WY path if available,
-    /// otherwise falls back to the sequential per-token gdn_decode loop.
+    /// / MTP verify). Picks the K=2/3/4, K∈{5..8} (wyN) or K=17 fused WY
+    /// path if available, otherwise falls back to the sequential per-token
+    /// gdn_decode loop.
     pub(super) fn decode_batched_conv_gdn(
         &self,
         ssm_state: &mut SsmLayerState,
@@ -337,105 +340,31 @@ impl Qwen3SsmLayer {
         } else if num_tokens == 17 && self.gdn_wy17_k.0 != 0 && wy17_enabled() {
             // ── K=17 (DFlash γ+1): fused WY-Chunkwise path ──
             //
-            // Conv epilogue: single fused launch over all 17 positions when
-            // the kernel is present (writes every rollback snapshot inline —
-            // replaces 17 conv launches + 17 copy_d2d, each serialized on the
-            // in-place conv_state update at 1-token occupancy). Requires the
-            // pool-contiguous intermediates layout (slot-major, stride
-            // conv_bytes — same contiguity the WY17 GDN call below already
-            // assumes for h_state_intermediates). Kill-switch:
-            // ATLAS_GDN_FUSED_CONV17=0 restores the per-token loop for A/B.
-            let conv_inter_base = ssm_state.conv_state_intermediates[0];
-            let inter_contiguous = ssm_state
-                .conv_state_intermediates
+            // Shared pool-layout arm (fused conv_kn epilogue + one wy17
+            // launch) — body lives in trait_decode_batched_conv_gdn_wyn.rs,
+            // dispatched identically for the chain-verify K∈{5..8} widths
+            // below.
+            self.decode_batched_conv_gdn_wyn(ssm_state, ctx, args, self.gdn_wy17_k)?;
+        } else if let Some(wyn_k) = self.wyn_kernel(num_tokens).filter(|_| {
+            // The wyN launch writes Hi_t at h_state_intermediates[0] +
+            // t*h_bytes — require the pool-contiguous layout it assumes
+            // (always true for ssm_pool slots); fail safe to the sequential
+            // fallback otherwise instead of corrupting memory.
+            let h_base = ssm_state.h_state_intermediates[0];
+            ssm_state
+                .h_state_intermediates
                 .iter()
-                .take(num_tokens)
+                .take(num_tokens - 1)
                 .enumerate()
-                .all(|(t, p)| p.0 == conv_inter_base.0 + (t * conv_bytes) as u64);
-            let fused_conv = self.gdn_verify_fused_conv_kn_k.0 != 0
-                && inter_contiguous
-                && !matches!(
-                    std::env::var("ATLAS_GDN_FUSED_CONV17").ok().as_deref(),
-                    Some("0")
-                );
-            if fused_conv {
-                ops::gdn_verify_fused_conv_kn(
-                    ctx.gpu,
-                    self.gdn_verify_fused_conv_kn_k,
-                    ssm_state.conv_state,
-                    deinterleaved,
-                    &self.ssm.conv1d,
-                    conv_out_buf,
-                    conv_inter_base,
-                    num_tokens as u32,
-                    conv_dim as u32,
-                    d_conv as u32,
-                    qk_ch,
-                    kd as u32,
-                    qkvz_size as u32, // input stride (BF16 elems between positions)
-                    conv_dim as u32,  // output stride (BF16 elems between positions)
-                    (conv_bytes / 4) as u32, // snapshot stride (FP32 elems)
-                    1e-6,
-                    stream,
-                )?;
-            } else {
-                for t in 0..(num_tokens as u32) {
-                    let qkv_t = deinterleaved.offset(t as usize * qkvz_size * bf16);
-                    let conv_out_t = conv_out_buf.offset(t as usize * conv_dim * bf16);
-                    ops::conv1d_update_l2norm(
-                        ctx.gpu,
-                        self.conv1d_l2norm_k,
-                        ssm_state.conv_state,
-                        qkv_t,
-                        &self.ssm.conv1d,
-                        conv_out_t,
-                        conv_dim as u32,
-                        d_conv as u32,
-                        1,
-                        qk_ch,
-                        kd as u32,
-                        1e-6,
-                        stream,
-                    )?;
-                    ctx.gpu.copy_d2d_async(
-                        ssm_state.conv_state,
-                        ssm_state.conv_state_intermediates[t as usize],
-                        conv_bytes,
-                        stream,
-                    )?;
-                }
-            }
-
-            let q_ptr = conv_out_buf;
-            let k_ptr = conv_out_buf.offset(key_dim * bf16);
-            let v_ptr = conv_out_buf.offset(key_dim * 2 * bf16);
-            let gate_ptr = gates_buf;
-            let beta_ptr = gates_buf.offset(nv * fp32);
-            let inter_stride_floats = (h_bytes / 4) as u32;
-            ops::gdn_decode_wy17(
-                ctx.gpu,
-                self.gdn_wy17_k,
-                ssm_state.h_state,
-                q_ptr,
-                k_ptr,
-                v_ptr,
-                gate_ptr,
-                beta_ptr,
-                gdn_out_buf,
-                ssm_state.h_state_intermediates[0],
-                inter_stride_floats,
-                1, // batch_size
-                nk as u32,
-                nv as u32,
-                kd as u32,
-                vd as u32,
-                conv_dim as u32, // qk_stride
-                conv_dim as u32, // v_stride
-                (nv * 2) as u32, // gb_stride
-                stream,
-            )?;
+                .all(|(t, p)| p.0 == h_base.0 + (t * h_bytes) as u64)
+        }) {
+            // ── K∈{5..8} chain verify: fused WY-Chunkwise path (wy5..wy8,
+            // one K-templated kernel source). Removes the serial per-token
+            // GDN fallback at these widths. Kill-switch: ATLAS_GDN_WYN=0. ──
+            self.decode_batched_conv_gdn_wyn(ssm_state, ctx, args, wyn_k)?;
         } else {
-            // ── K!=2,17: sequential per-token path ──
+            // ── No fused arm (K>17, K∈{5..8} with wyN absent/killed, or
+            // non-pool intermediates): sequential per-token path ──
             //
             // gated_delta_rule_decode expects FP32 Q/K/V (see kernel signature),
             // but causal_conv1d_update_l2norm outputs BF16 by default. Reading
